fixed ticked 4123 and 4124

+++ pandora_console/godmode/agentes/module_manager_editor_common.php
@@ -100,6 +100,7 @@ global $__code_from;
 
 $disabledBecauseInPolicy = false;
 $disabledTextBecauseInPolicy = '';
+$largeClassDisabledBecauseInPolicy = '';
 
 $page = get_parameter('page', '');
 if (strstr($page, "policy_modules") === false && $id_agent_module) {
@@ -108,7 +109,14 @@ if (strstr($page, "policy_modules") === false && $id_agent_module) {
 	else
 		$disabledBecauseInPolicy = false;
 	if ($disabledBecauseInPolicy)
-		$disabledTextBecauseInPolicy = 'disabled = "disabled"';
+		$disabledTextBecauseInPolicy = 'readonly = "readonly"';
+}
+
+if($disabledBecauseInPolicy){
+	$classdisabledBecauseInPolicy = 'readonly';
+	$largeClassDisabledBecauseInPolicy = 'class=readonly';
+} else {
+	$classdisabledBecauseInPolicy = '';
 }
 
 $update_module_id = (int) get_parameter_get ('update_module');
@@ -129,7 +137,7 @@ $table_simple->colspan[6][1] = 3;
 
 $table_simple->data[0][0] = __('Name');
 $table_simple->data[0][1] = html_print_input_text_extended ('name',
-	io_safe_output($name), 'text-name', '', 45, 100, $disabledBecauseInPolicy, '', 'autocomplete="off"', true);
+	io_safe_output($name), 'text-name', '', 45, 100, $disabledBecauseInPolicy, '', $largeClassDisabledBecauseInPolicy, true);
 //$table_simple->data[0][1] = html_print_input_text ('name',
 //	io_safe_output($name), '', 45, 100, true, $disabledBecauseInPolicy);
 
@@ -236,54 +244,56 @@ $table_simple->data[2][1] = '';
 if (!modules_is_string_type($id_module_type) || $edit) {
 	$table_simple->data[2][1] .= '<span id="minmax_warning"><em>'.__('Min. ').'</em>';
 	$table_simple->data[2][1] .= html_print_input_text ('min_warning', $min_warning,
-		'', 10, 255, true, $disabledBecauseInPolicy);
+		'', 10, 255, true, $disabledBecauseInPolicy, false, '', $classdisabledBecauseInPolicy);
 	$table_simple->data[2][1] .= '<br /><em>'.__('Max.').'</em>';
 	$table_simple->data[2][1] .= html_print_input_text ('max_warning', $max_warning,
-		'', 10, 255, true, $disabledBecauseInPolicy).'</span>';
+		'', 10, 255, true, $disabledBecauseInPolicy, false, '', $classdisabledBecauseInPolicy).'</span>';
 }
 if (modules_is_string_type($id_module_type) || $edit) {
 	$table_simple->data[2][1] .= '<span id="string_warning"><em>'.__('Str.').'</em>';
 	$table_simple->data[2][1] .= html_print_input_text ('str_warning', $str_warning, 
-		'', 10, 255, true, $disabledBecauseInPolicy).'</span>';
+		'', 10, 255, true, $disabledBecauseInPolicy, false, '', $classdisabledBecauseInPolicy).'</span>';
 }
 
 $table_simple->data[2][1] .= '<br /><em>'.__('Inverse interval').'</em>';
-$table_simple->data[2][1] .= html_print_checkbox ("warning_inverse", 1,
-	$warning_inverse, true);
+$table_simple->data[2][1] .= html_print_checkbox ("warning_inverse", 1, $warning_inverse, true, $disabledBecauseInPolicy);
 $table_simple->data[2][2] = __('Critical status').' ' . ui_print_help_icon ('critical_status', true);
 $table_simple->data[2][3] = '';
 
 if (!modules_is_string_type($id_module_type) || $edit) {
 	$table_simple->data[2][3] .= '<span id="minmax_critical"><em>'.__('Min. ').'</em>';
 	$table_simple->data[2][3] .= html_print_input_text ('min_critical', $min_critical,
-		'', 10, 255, true, $disabledBecauseInPolicy);
+		'', 10, 255, true, $disabledBecauseInPolicy, false, '', $classdisabledBecauseInPolicy);
 	$table_simple->data[2][3] .= '<br /><em>'.__('Max.').'</em>';
 	$table_simple->data[2][3] .= html_print_input_text ('max_critical', $max_critical,
-		'', 10, 255, true, $disabledBecauseInPolicy).'</span>';
+		'', 10, 255, true, $disabledBecauseInPolicy, false, '', $classdisabledBecauseInPolicy).'</span>';
 }
 if (modules_is_string_type($id_module_type) || $edit) {
 	$table_simple->data[2][3] .= '<span id="string_critical"><em>'.__('Str.').'</em>';
 	$table_simple->data[2][3] .= html_print_input_text ('str_critical', $str_critical, 
-		'', 10, 255, true, $disabledBecauseInPolicy).'</span>';
+		'', 10, 255, true, $disabledBecauseInPolicy, false, '', $classdisabledBecauseInPolicy).'</span>';
 }
 
 $table_simple->data[2][3] .= '<br /><em>'.__('Inverse interval').'</em>';
-$table_simple->data[2][3] .= html_print_checkbox ("critical_inverse", 1, $critical_inverse, true);
+$table_simple->data[2][3] .= html_print_checkbox ("critical_inverse", 1, $critical_inverse, true, $disabledBecauseInPolicy);
 
 /* FF stands for Flip-flop */
 $table_simple->data[3][0] = __('FF threshold').' ' . ui_print_help_icon ('ff_threshold', true);
 $table_simple->colspan[3][1] = 3;
 
-$table_simple->data[3][1] = html_print_radio_button ('each_ff', 0, '', $each_ff, true) . ' ' . __('All state changing') . ' : ';
-$table_simple->data[3][1] .= html_print_input_text ('ff_event', $ff_event, '', 5
-, 15, true, $disabledBecauseInPolicy) . '<br />';
-$table_simple->data[3][1] .= html_print_radio_button ('each_ff', 1, '', $each_ff, true) . ' ' . __('Each state changing') . ' : ';
+$table_simple->data[3][1] = html_print_radio_button ('each_ff', 0, '', $each_ff, true, $disabledBecauseInPolicy) . ' ' . __('All state changing') . ' : ';
+$table_simple->data[3][1] .= html_print_input_text ('ff_event', $ff_event, '', 5, 15, true,
+													$disabledBecauseInPolicy, false, '', $classdisabledBecauseInPolicy) . '<br />';
+$table_simple->data[3][1] .= html_print_radio_button ('each_ff', 1, '', $each_ff, true, $disabledBecauseInPolicy) . ' ' . __('Each state changing') . ' : ';
 $table_simple->data[3][1] .= __('To normal');
-$table_simple->data[3][1] .= html_print_input_text ('ff_event_normal', $ff_event_normal, '', 5, 15, true, $disabledBecauseInPolicy) . ' ';
+$table_simple->data[3][1] .= html_print_input_text ('ff_event_normal', $ff_event_normal, '', 5, 15, true,
+													$disabledBecauseInPolicy, false, '', $classdisabledBecauseInPolicy) . ' ';
 $table_simple->data[3][1] .= __('To warning');
-$table_simple->data[3][1] .= html_print_input_text ('ff_event_warning', $ff_event_warning, '', 5, 15, true, $disabledBecauseInPolicy) . ' ';
+$table_simple->data[3][1] .= html_print_input_text ('ff_event_warning', $ff_event_warning, '', 5, 15, true,
+													$disabledBecauseInPolicy, false, '', $classdisabledBecauseInPolicy) . ' ';
 $table_simple->data[3][1] .= __('To critical');
-$table_simple->data[3][1] .= html_print_input_text ('ff_event_critical', $ff_event_critical, '', 5, 15, true, $disabledBecauseInPolicy);
+$table_simple->data[3][1] .= html_print_input_text ('ff_event_critical', $ff_event_critical, '', 5, 15, true,
+													$disabledBecauseInPolicy, false, '', $classdisabledBecauseInPolicy);
 $table_simple->data[4][0] = __('Historical data');
 if($disabledBecauseInPolicy) {
 	// If is disabled, we send a hidden in his place and print a false checkbox because HTML dont send disabled fields and could be disabled by error
@@ -309,16 +319,16 @@ $table_advanced->colspan = array ();
 $table_advanced->data[0][0] = __('Description');
 $table_advanced->colspan[0][1] = 6;
 $table_advanced->data[0][1] = html_print_textarea ('description', 2, 65,
-	$description, $disabledTextBecauseInPolicy, true);
+	$description, $disabledTextBecauseInPolicy, true, $largeClassDisabledBecauseInPolicy);
 
 $table_advanced->data[1][0] = __('Custom ID');
 $table_advanced->colspan[1][1] = 2;
 $table_advanced->data[1][1] = html_print_input_text ('custom_id', $custom_id,
-	'', 20, 65, true);
+	'', 20, 65, true, $disabledBecauseInPolicy, false, '', $classdisabledBecauseInPolicy);
 
 $table_advanced->data[1][3] = __('Unit');
 $table_advanced->data[1][4] = html_print_input_text ('unit', $unit,
-	'', 20, 65, true);
+	'', 20, 65, true, $disabledBecauseInPolicy, false, '', $classdisabledBecauseInPolicy);
 $table_advanced->colspan[1][4] = 3;
 
 $module_id_policy_module = 0;
@@ -357,7 +367,7 @@ if ($moduletype == MODULE_DATA) {
 else {
 	$table_advanced->data[2][0] = __('Interval') . ui_print_help_icon ('module_interval', true);
 	$table_advanced->colspan[2][1] = 2;
-	$table_advanced->data[2][1] = html_print_extended_select_for_time ('module_interval' , $interval, '', '', '0', false, true, false, false);
+	$table_advanced->data[2][1] = html_print_extended_select_for_time ('module_interval' , $interval, '', '', '0', false, true, false, false, $classdisabledBecauseInPolicy, $disabledBecauseInPolicy);
 }
 
 $table_advanced->data[2][1] .= html_print_input_hidden ('moduletype', $moduletype, true);
@@ -372,9 +382,9 @@ $table_advanced->colspan[2][4] = 3;
 $table_advanced->data[3][0] = __('Min. Value');
 $table_advanced->colspan[3][1] = 2;
 
-$table_advanced->data[3][1] = html_print_input_text ('min', $min, '', 5, 15, true, $disabledBecauseInPolicy). ' ' . ui_print_help_tip (__('Any value below this number is discarted.'), true);
+$table_advanced->data[3][1] = html_print_input_text ('min', $min, '', 5, 15, true, $disabledBecauseInPolicy, false, '', $classdisabledBecauseInPolicy). ' ' . ui_print_help_tip (__('Any value below this number is discarted.'), true);
 $table_advanced->data[3][3] = __('Max. Value');
-$table_advanced->data[3][4] = html_print_input_text ('max', $max, '', 5, 15, true, $disabledBecauseInPolicy). ' ' . ui_print_help_tip (__('Any value over this number is discarted.'), true);
+$table_advanced->data[3][4] = html_print_input_text ('max', $max, '', 5, 15, true, $disabledBecauseInPolicy, false, '', $classdisabledBecauseInPolicy). ' ' . ui_print_help_tip (__('Any value over this number is discarted.'), true);
 $table_advanced->colspan[3][4] = 3;
 
 $table_advanced->data[4][0] = __('Export target');
@@ -403,12 +413,12 @@ else {
 }
 $table_advanced->data[4][3] = __('Throw unknown events');
 $table_advanced->data[4][4] = html_print_checkbox('throw_unknown_events',
-	1, $throw_unknown_events_check, true);
+	1, $throw_unknown_events_check, true, $disabledBecauseInPolicy);
 $table_advanced->colspan[4][4] = 3;
 
 $table_advanced->data[5][0] = __('FF interval') . ' ' . ui_print_help_icon ('ff_interval', true);
 $table_advanced->data[5][1] = html_print_input_text ('module_ff_interval', $ff_interval,
-	'', 5, 10, true, $disabledBecauseInPolicy) .
+	'', 5, 10, true, $disabledBecauseInPolicy, false, '', $classdisabledBecauseInPolicy) .
 	ui_print_help_tip (__('Module execution flip flop time interval (in secs).'), true);
 $table_advanced->colspan[5][1] = 2;
 
@@ -418,7 +428,7 @@ $module_type_name = modules_get_type_name($id_module_type);
 $table_advanced->data[5][4] = '';
 if (preg_match ('/async/', $module_type_name) || $edit) {
 	$table_advanced->data[5][4] .= '<span id="ff_timeout">' . html_print_input_text ('ff_timeout', $ff_timeout,
-		'', 5, 10, true, $disabledBecauseInPolicy).ui_print_help_tip (__('Timeout in secs from start of flip flop counting. If this value is exceeded, FF counter is reset. Set to 0 for no timeout.'), true) . '</span>';
+		'', 5, 10, true, $disabledBecauseInPolicy, false, '', $classdisabledBecauseInPolicy).ui_print_help_tip (__('Timeout in secs from start of flip flop counting. If this value is exceeded, FF counter is reset. Set to 0 for no timeout.'), true) . '</span>';
 }
 if (!preg_match ('/async/', $module_type_name) || $edit) {
 	$table_advanced->data[5][4] .= '<span id="ff_timeout_disable">' . __('Disabled') . ui_print_help_tip (__('This value can be set only in the async modules.'), true) . '</span>';
@@ -460,7 +470,7 @@ if (!tags_has_user_acl_tags($config["id_user"])) {
 			FROM ttag a, $__table_modules b 
 			WHERE a.id_tag = b.id_tag AND $__id_where = $__id )
 			ORDER BY name", 'id_tag_available[]', '', '','','',
-		true, true, false, false, 'width: 200px', '5');
+		true, true, false, $disabledBecauseInPolicy, 'width: 200px', '5');
 }
 else {
 	$user_tags = tags_get_user_tags($config["id_user"], "AW");
@@ -476,7 +486,7 @@ else {
 				FROM ttag a, $__table_modules b 
 				WHERE a.id_tag = b.id_tag AND $__id_where = $__id )
 				ORDER BY name", 'id_tag_available[]', '', '','','',
-			true, true, false, false, 'width: 200px', '5');
+			true, true, false, $disabledBecauseInPolicy, 'width: 200px', '5');
 	}
 	else {
 		$table_advanced->data[6][1] = html_print_select_from_sql (
@@ -487,7 +497,7 @@ else {
 				FROM ttag a, $__table_modules b 
 				WHERE a.id_tag = b.id_tag AND $__id_where = $__id )
 				ORDER BY name", 'id_tag_available[]', '', '','','',
-			true, true, false, false, 'width: 200px', '5');
+			true, true, false, $disabledBecauseInPolicy, 'width: 200px', '5');
 	}
 }
 $table_advanced->data[6][2] = html_print_image('images/darrowright.png', true, array('id' => 'right', 'title' => __('Add tags to module'))); //html_print_input_image ('add', 'images/darrowright.png', 1, '', true, array ('title' => __('Add tags to module')));
@@ -520,20 +530,20 @@ $table_advanced->data[7][0] .= ui_print_help_tip(
 	__('The module still stores data but the alerts and events will be stop'), true);
 $table_advanced->colspan[7][1] = 7;
 $table_advanced->data[7][1] = html_print_checkbox('quiet_module', 1,
-	$quiet_module, true);
+	$quiet_module, true, $disabledBecauseInPolicy);
 
 $table_advanced->data[8][0] = __('Critical instructions') .
 	ui_print_help_tip(__("Instructions when the status is critical"), true);
-$table_advanced->data[8][1] = html_print_textarea ('critical_instructions', 2, 65, $critical_instructions, '', true);
+$table_advanced->data[8][1] = html_print_textarea ('critical_instructions', 2, 65, $critical_instructions, $disabledTextBecauseInPolicy, true, $largeClassDisabledBecauseInPolicy);
 $table_advanced->colspan[8][1] = 6;
 
 $table_advanced->data[9][0] = __('Warning instructions') .
 	ui_print_help_tip(__("Instructions when the status is warning"), true);
-$table_advanced->data[9][1] = html_print_textarea ('warning_instructions', 2, 65, $warning_instructions, '', true);
+$table_advanced->data[9][1] = html_print_textarea ('warning_instructions', 2, 65, $warning_instructions, $disabledTextBecauseInPolicy, true, $largeClassDisabledBecauseInPolicy);
 $table_advanced->colspan[9][1] = 6;
 
 $table_advanced->data[10][0] = __('Unknown instructions'). ui_print_help_tip(__("Instructions when the status is unknown"), true);
-$table_advanced->data[10][1] = html_print_textarea ('unknown_instructions', 2, 65, $unknown_instructions, '', true);
+$table_advanced->data[10][1] = html_print_textarea ('unknown_instructions', 2, 65, $unknown_instructions, $disabledTextBecauseInPolicy, true, $largeClassDisabledBecauseInPolicy);
 $table_advanced->colspan[10][1] = 6;
 
 if (isset($id_agente) && $moduletype == MODULE_DATA) {
@@ -545,21 +555,21 @@ if (isset($id_agente) && $moduletype == MODULE_DATA) {
 else {
 	$table_advanced->data[11][0] = __('Cron') .
 		ui_print_help_tip (__('If cron is set the module interval is ignored and the module runs on the specified date and time'), true);
-	$table_advanced->data[11][1] = html_print_extended_select_for_cron ($hour, $minute, $mday, $month, $wday, true, false);
+	$table_advanced->data[11][1] = html_print_extended_select_for_cron ($hour, $minute, $mday, $month, $wday, true, $disabledBecauseInPolicy);
 	$table_advanced->colspan[11][1] = 6;
 }
 
 $table_advanced->data[12][0] = __('Timeout');
-$table_advanced->data[12][1] = html_print_input_text ('max_timeout', $max_timeout, '', 5, 10, true). ' ' . ui_print_help_tip (__('Seconds that agent will wait for the execution of the module.'), true);
+$table_advanced->data[12][1] = html_print_input_text ('max_timeout', $max_timeout, '', 5, 10, true, $disabledBecauseInPolicy, false, '', $classdisabledBecauseInPolicy). ' ' . ui_print_help_tip (__('Seconds that agent will wait for the execution of the module.'), true);
 $table_advanced->data[12][2] = '';
 $table_advanced->data[12][3] = __('Retries');
-$table_advanced->data[12][4] = html_print_input_text ('max_retries', $max_retries, '', 5, 10, true). ' ' . ui_print_help_tip (__('Number of retries that the module will attempt to run.'), true);
+$table_advanced->data[12][4] = html_print_input_text ('max_retries', $max_retries, '', 5, 10, true, $disabledBecauseInPolicy, false, '', $classdisabledBecauseInPolicy). ' ' . ui_print_help_tip (__('Number of retries that the module will attempt to run.'), true);
 $table_advanced->colspan[12][4] = 3;
 
 if (check_acl ($config['id_user'], 0, "PM")) {
 	$table_advanced->data[13][0] = __('Category');
 	$table_advanced->data[13][1] = html_print_select(
-		categories_get_all_categories('forselect'), 'id_category', $id_category, '', __('None'), 0, true);
+		categories_get_all_categories('forselect'), 'id_category', $id_category, '', __('None'), 0, true, false, true, "", $disabledBecauseInPolicy);
 	$table_advanced->colspan[13][1] = 6;
 }
 else {
@@ -584,16 +594,20 @@ if (isset($module_macros)) {
 	if (is_array($module_macros)) {
 		foreach ($module_macros as $macro_name => $macro_value) {
 			$table_macros->data[$macro_count][0] = __('Name');
-			$table_macros->data[$macro_count][1] = html_print_input_text ('module_macro_names[]', $macro_name, '', 50, 60, true);
+			$table_macros->data[$macro_count][1] = html_print_input_text ('module_macro_names[]', $macro_name, '', 50, 60, true, $disabledBecauseInPolicy, false, '', $classdisabledBecauseInPolicy);
 			$table_macros->data[$macro_count][2] = __('Value');
-			$table_macros->data[$macro_count][3] = html_print_input_text ('module_macro_values[]', $macro_value, '', 50, 60, true);
-			$table_macros->data[$macro_count][4] = '<a href="javascript: delete_macro(' . $macro_count . ');">' . html_print_image('images/cross.png', true) . '</a>';
+			$table_macros->data[$macro_count][3] = html_print_input_text ('module_macro_values[]', $macro_value, '', 50, 60, true, $disabledBecauseInPolicy, false, '', $classdisabledBecauseInPolicy);
+			if(!$disabledBecauseInPolicy){
+				$table_macros->data[$macro_count][4] = '<a href="javascript: delete_macro(' . $macro_count . ');">' . html_print_image('images/cross.png', true) . '</a>';
+			}
 			$macro_count++;
 		}
 	}
 }
-$table_macros->data[$macro_count][0] = '<span>'.__('Custom macros').'</span> <a href="javascript:add_macro();">'.html_print_image('images/add.png',true).'</a>';
-$table_macros->colspan[$macro_count][0] = 5;
+if(!$disabledBecauseInPolicy){
+	$table_macros->data[$macro_count][0] = '<span>'.__('Custom macros').'</span> <a href="javascript:add_macro();">'.html_print_image('images/add.png',true).'</a>';
+	$table_macros->colspan[$macro_count][0] = 5;
+}
 $macro_count++;
 
 html_print_input_hidden ('module_macro_count', $macro_count);
